pyproject.toml
[project]
name = "travelplanner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.136.0",
    "langchain>=1.2.15",
    "langchain-anthropic>=1.4.1",
    "langchain-community>=0.4.1",
    "langchain-google-community>=3.0.5",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.15",
    "langchain-tavily>=0.2.18",
    "langgraph>=1.1.9",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "streamlit>=1.56.0",
    "uvicorn>=0.45.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "ruff>=0.15.11",
]
